+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this package are documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-03-30
+
+### Added
+- Launch options:
+	- -scene "sceneName"
+	- -server
+	- -serverCam
+	- -ip "address"
+	- -port "port"
+
+### Changed
+- Network component has dirty built in
+- Network components have unique ids to prevent mixed messages on game objects with separate network scripts.
+
 ## [0.1.1] - 2021-03-30
 
 ### Changed